[BE] refactor: subproject 위치 변경
